Flask_Server/requirements.txt
absl-py==2.1.0
alembic==1.13.1
aniso8601==9.0.1
anyio==4.3.0
astor==0.8.1
asttokens==2.4.0
astunparse==1.6.3
attrdict==2.0.1
Babel==2.15.0
backcall==0.2.0
bce-python-sdk==0.9.7
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.1.1
cssselect==1.2.0
cssutils==2.10.2
cycler==0.12.1
Cython==3.0.10
debugpy==1.8.0
decorator==5.1.1
dill==0.3.8
dnspython==2.6.1
et-xmlfile==1.1.0
eventlet==0.36.1
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.14.0
fire==0.6.0
Flask==3.0.3
flask-babel==4.0.0
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
future==1.0.0
gast==0.4.0
gdown==5.1.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.1
imgaug==0.4.0
importlib-metadata==7.1.0
importlib-resources==6.4.0
ipykernel==6.29.0
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter-client==8.6.0
jupyter-core==5.7.2
keras==2.13.1
kiwisolver==1.4.5
lazy-loader==0.4
libclang==18.1.1
lmdb==1.4.1
lxml==5.2.1
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==24.0
paddleocr==2.7.3
paddlepaddle-gpu==2.6.1
pandas==2.0.3
parso==0.8.4
pdf2docx==0.5.8
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.0
premailer==3.10.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pyclipper==1.3.0.post5
pycryptodome==3.20.0
Pygments==2.18.0
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==26.0.2
rapidfuzz==3.9.0
rarfile==4.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.13.2
shapely==2.0.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
thop==0.1.1.post2209072238
tifffile==2023.7.10
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.2
triton==2.3.0
typing-extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.11
urllib3==2.2.1
visualdl==2.5.3
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
zipp==3.18.1
